Refactor Gender
