+++ src/web/app/pages-help/instructor-help-page/instructor-help-students-section/instructor-help-students-section.component.spec.ts
@@ -3,8 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Component, Input } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { JoinState, Student } from '../../../../types/api-output';
-import { Gender } from '../../../../types/gender';
+import { Gender, JoinState, Student } from '../../../../types/api-output';
 import { SearchStudentsTable } from '../../../pages-instructor/instructor-search-page/instructor-search-page.component';
 import { StudentProfile } from '../../../pages-instructor/student-profile/student-profile';
 import { InstructorHelpStudentsSectionComponent } from './instructor-help-students-section.component';
