+++ src/web/app/pages-student/student-profile-page/student-profile-gender.pipe.ts
@@ -1,5 +1,6 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { Gender } from '../../../types/gender';
+import { Gender } from '../../../types/api-output';
+
 
 /**
  * Pipe to handle the display of gender options in the student profile page.
