--- src/web/types/gender.ts
@@ -1,20 +1,0 @@
-/**
- * Represents the gender of a student.
- */
-export enum Gender {
-
-  /**
-   * Male gender.
-   */
-  MALE = 'MALE',
-
-  /**
-   * Female gender.
-   */
-  FEMALE = 'FEMALE',
-
-  /**
-   * Not specified gender.
-   */
-  OTHER = 'OTHER',
-}
